[TESTING] pruning and updating tests, only forge data are tested. there was too many tests that didnt work properly before

+++ RAMSIS/cli/model.py
@@ -1,4 +1,6 @@
 import typer
+from rich import print
+from rich.table import Table
 from marshmallow import EXCLUDE
 from RAMSIS.db import db_url, session_handler, init_db
 from ramsis.datamodel import ModelConfig
@@ -13,6 +15,23 @@ app = typer.Typer()
 
 
 @app.command()
+def ls(help="Outputs list of models"):
+    with session_handler(db_url) as session:
+        model_configs = session.execute(
+            select(ModelConfig)).scalars().all()
+        for model in model_configs:
+            table = Table(show_footer=False,
+                          title=f"Model Config {model.name}",
+                          title_justify="left")
+            table.add_column("attribute")
+            table.add_column("value")
+            for attr in ModelConfig.__table__.columns:
+                table.add_row(str(attr.name), str(getattr(model, attr.name)))
+
+            print(table)
+
+
+@app.command()
 def delete(
         model_name: str,
         force: bool = typer.Option(
@@ -23,17 +42,17 @@ def delete(
                 name=model_name)).\
             scalar_one_or_none()
         if not model_config:
-            typer.echo("Model does not exist")
+            print("Model does not exist")
             raise typer.Exit()
         if model_config.runs:
-            typer.echo("Model config is associated with model runs, "
-                       "cannot delete. You can disable instead.")
+            print("Model config is associated with model runs, "
+                  "cannot delete. You can disable instead.")
             raise typer.Exit()
             if not force:
                 delete = typer.confirm("Are you sure you want to delete the  "
                                        f"model with name: {model_name}?")
                 if not delete:
-                    typer.echo("Not deleting")
+                    print("Not deleting")
                     raise typer.Abort()
         session.delete(model_config)
         session.commit()
@@ -51,7 +70,7 @@ def load(
     if success:
         pass
     else:
-        typer.echo(f"Error, db could not be initialized: {success}")
+        print(f"Error, db could not be initialized: {success}")
         raise typer.Exit()
     with session_handler(db_url) as session:
         with open(model_config, "r") as model_read:
@@ -64,16 +83,16 @@ def load(
                 scalar_one_or_none()
 
             if existing_config:
-                typer.echo(f"Model config exists for {existing_config.name}.")
+                print(f"Model config exists for {existing_config.name}.")
                 # Check if there are already completed forecasts
                 # associated before allowing modification.
-                runs_exist = True if existing_config.runs else False
-                if runs_exist:
+                print(existing_config.runs, "existing runs")
+                if existing_config.runs:
                     session.rollback()
-                    typer.echo("Model runs already exist for this config"
-                               " Please upload the config with a new name.")
-                    typer.Exit(code=1)
-                typer.echo("deleting existing config")
+                    print("Model runs already exist for this config"
+                          " Please upload the config with a new name.")
+                    raise typer.Exit(code=1)
+                print("deleting existing config")
                 session.delete(existing_config)
                 new_model_config = ModelConfigurationSchema(
                     unknown=EXCLUDE, context={"session": session}).load(config)
@@ -81,10 +100,10 @@ def load(
             else:
                 new_model_config = ModelConfigurationSchema(
                     unknown=EXCLUDE, context={"session": session}).load(config)
-                typer.echo("Model config is being added for "
-                           f"{new_model_config.name}")
+                print("Model config is being added for "
+                      f"{new_model_config.name}")
 
             session.add(new_model_config)
             session.commit()
-            typer.echo("A model has been configured with the name: "
-                       f"{new_model_config.name}, id: {new_model_config.id}")
+            print("A model has been configured with the name: "
+                  f"{new_model_config.name}, id: {new_model_config.id}")
